requirements.txt
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
authlib==1.3.0
python-dotenv==1.0.0
APScheduler==3.10.4
psycopg2-binary==2.9.9
